Add socket backlog metric
If all the workers are busy or max connections is reached, new
connections will queue in the socket backlog, which defaults to 2048.
The `gunicorn.backlog` metric provide visibility into this queue, and
give an idea on concurrency, and worker saturation.

This also adds a distinction between the `timer` and `histogram` statsd
metric types, which although treated the same, can be difference, for
e.g. in this case histogram is not a timer: https://github.com/b/statsd_spec#timers